CI: Fix job names to differentiate
The hacs and hassfest validation jobs were both titled 'validate' which
made it impossible to use a hard validation on just one of them in
branch protections. As such, we're differentiating some so that we can
utilize branch protection but also so that it's easier to tell which
action failed.

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>