Fix: Make options updates update the stored config
The options flow stores changes in an options sub-hash of the entry
object. This means that by default changes to a configured entry aren't
always reflected.

To resolve this, we need to attach an update listener that guards
against the updates that it is making so that the config gets fully
updated.

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>