Feat: Add event prefixes
If more than one Rental Control is defined in a system it can be
difficult to differentiate the calendar events in some cases. Give the
end user a way to force a differentiation by allowing them to define a
prefix that gets pre-pended to all event summaries.

ie: AirBnB has all reservations as 'Reserved' if more than one unit is
defined it is useful to have those reservations differentiated such as:

'Unit 1 - Reserved'
'Unit 2 - Reserved'

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>