Feat: Allow integration to work without Keymaster
There has been a request to allow the integration to not depend on
Keymaster. As we are not requiring a lock to actually be used, this
makes sense.

However, we will lose functionality that is planned for the future
related to modifying checkin/checkout times of individual events if
there is no door code slot available as the intent is to inform the
integration of changes by way of modification to the in/out times of the
assigned door code.

Issue: #45
Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>