Fix: Correct error strings
The error string for a URL that returns bad data incorrectly states that
AirBnB is the only supported calendar. This is incorrect as we now
verify that the data returns is of type text/calendar and accept any
calendar that passes this.

Additionally, the error string for invalid URLs was not giving a reason
why it was considered invalid.

Issue: #73
Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>