Fix: Better URL error handling
Additional testing found that during the calendar validation some errors
were getting overwritten by other errors and thus clear messaging to the
end user was not happening.

In particular, if a calendar platform returned a non-200 code this would
present as an issue of invalid calendar data (technically accurate). It
should, however have presented as an unknown error.

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>