Fix: Update HACS to do zip releases
While doing some looking into analytics it was discovered that all
versions of Rental Control were reporting version 0.0.0 which means that
all installed versions have been installing directly from the
repository. This makes it hard to understand what the install base is!

As zip releases were supposed to be on for a long time this actually
fixes it to have them on.

It also fixes the fact that the calendar domain was not listed

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>