Feat: Configurable calendar refresh frequency
Some calendars have a limited number of times per day that the calendar
can be refreshed. Since the refresh was originally hard coded at no more
than once every 2 minutes this caused polling exhaustion for some users.

The configuration now supports a polling frequency between as often as
possible (roughly every 30 seconds because of the HA event machine) up
to once per day. This configuration is given as a number of minutes
between 0 and 1440.

Issue: #79
Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>