Feat: Add code generators
Teach the event sensor 3 ways to generate door codes.

The default (and fall back) is check-in/out day based where the check-in
day and check-out day are combined to create a 4 digit code

A random 4 digit code seeded from the event description

Finally, the last 4 digits of a phone number - only works if the
description has the text 'Last 4 Digits' followed quickly by a 4 digit
number. This is by far the most stable, but only works if the conditions
are correct

Issue: #38
Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>