Fix: Remove setting slot notification
Not everyone is going to want slot notifications. It's also sorta creepy
given that it can be seen as an invasion of privacy. As such, this
integration will not be turning it on. There may be a feature later to
add an additional configuration knob for if notifications should be
enabled or not.

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>