Fix: Prevent inaccessible calendars from deleting
A calendar that is inaccessible when the integration is starting up
primarily during a restart of HA, but also when the integration is
reloaded, could potentially cause currently defined lock slots to get
deleted. This can cause major issues, particularly if the slot had an
override on the code in place as the override is lost.

Instead, we keep the sensors as unavailable and do not fire any events
related to setting or clearing the the codes until we have managed to
load the calendar successfully.

Issue: Fixes #126
Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>