Fix: Stringify rental_control_name in automations
If a RC is defined with a name that is purely numeric then the default
behavior of yaml is to treat an numeric as a numeric and not a string.
This has a side effect of not properly stringifing the name where it
needs to be a string in the automations.

Fixes: #164
Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>