Refactor: Start migration of event overrides
Move the event overrides logic to it's own object.

This makes the code in RentalControl itself easier to read and also
centralizes most of the logic.

The majority of the logic is all sync instead of async code to help make
sure we're only operating on a given state at a time

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>