Chore: Remove mapping sensor creation
The MappingSensor has never really worked particularly well. As it never
properly worked, all the automations that were expected to start using
it never did.

Additionally, since the shipped automations are going away, the need for
the sensor is even less relevant.

Remove the loading of it.

Also, fix issues with reload of the integration

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>