Feat: Add should_update_code to configuration
Add new flag for data based codes to update the code if the start or end
date of future events change. This will allow for the code to be updated
if the event changes and the code is no longer valid.

Schema version is updated to 7 with the flag defaulting to False for
existing configurations and True for new configurations.

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>