Chore: Force update of requirements
The previous version of icalendar is no longer in sync with the version
that is used by core Home Assistant. Make sure that installing the
integration uses the correct version and also forces an upgrade of
x-wr-timezone to the latest version.

Finally, since HA 2025.1.0 is the version that explicitly starts using
icalendar 6.1.0, make sure that is the base version needed.

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>