CI: Update configuration for pre-commit CI
Make sure that pre-commit-ci is adding a DCO line. Upstream does not
view this as something that they should directly support or even do, but
since it is possible to set more than just the subject line, we can
effectively add the DCO line to the commit message just at the top of
the body instead of the footer.

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>