Fix: Fix calendar loading error handling
Airbnb's calendar sometimes returns a 200 OK response with an empty
body. This is a valid response, but it should be handled gracefully.
If the calendar is empty, we should not set the calendar to an empty
list, but rather keep the old calendar if the calendar has more than
one event. This prevents the calendar from being set to an empty list
when it should not be.

CAUTION: Calendars that truly purged all events may still end up having
events when they should not and the only way to fix this is to restart
the integration.

Issue: Fixes #245
Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>