Refactor!: Update to work with Keymaster v0.1.0-b0
Home Assistant 2025.8.0 introduced changes in how ZWave operates in a
way that causes Keymaster < v0.1.0-b0 to not work properly.

As such, this update requires Keymaster to be at least v0.1.0-b0
to function properly.

This is a breaking change, as it requires the user to update
Keymaster to the latest version as such it is incompatible with
Keymaster < v0.1.0-b0.

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>